pyproject.toml
[project]
name = "python-fastapi-admin"
version = "0.1.0"
description = "python fastapi admin"
authors = [
    {name = "lsugarj"}
]

readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "uvicorn (>=0.42.0,<0.43.0)",
    "fastapi (>=0.135.2,<0.136.0)",
    "asyncmy (>=0.2.11,<0.3.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (==4.0.1)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "sqlmodel (>=0.0.37,<0.0.38)",
    "redis (>=7.4.0,<8.0.0)",
    "python-json-logger (>=4.1.0,<5.0.0)",
    "prometheus-client (>=0.24.1,<0.25.0)"
]

[tool.poetry]
packages = [
    { include = "app", from = "." }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
